iteration on setting up containerinspection component
